apps/python-backend/requirements.txt
# Optimized requirements with specific versions for better caching
fastapi==0.104.1
numpy>=1.26.0
uvicorn[standard]==0.24.0
Pillow>=10.1.0
python-multipart==0.0.20

# PyTorch CPU-only for faster builds (CLIP dependency) - Python 3.12 compatible
torch>=2.2.0+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.17.0+cpu --index-url https://download.pytorch.org/whl/cpu

# CLIP dependencies
ftfy
regex
tqdm

# Use openai-clip package instead (more up-to-date)
openai-clip
